refactor: move rgbaToHex to shared @notra/utils/color
