+++ apps/console/src/lib/integrations/form.ts
@@ -126,17 +126,6 @@ export function getInitialApiKeyStyle(server?: McpServer): ApiKeyStyle {
   return hasStoredBearerHeader(server) ? "bearer" : "headers";
 }
 
-const HEX_CHANNEL_MAX = 255;
-
-export function rgbaToHex(channels: readonly number[]) {
-  const toHex = (channel: number | undefined) =>
-    Math.round(Math.min(Math.max(channel ?? 0, 0), HEX_CHANNEL_MAX))
-      .toString(16)
-      .padStart(2, "0");
-
-  return `#${toHex(channels[0])}${toHex(channels[1])}${toHex(channels[2])}`;
-}
-
 export function getIntegrationInitials(name: string) {
   return name.trim().slice(0, 2).toUpperCase() || "?";
 }
